Some type modifications in DataViewPrototype

+++ libraries/script-engine/src/DataViewPrototype.h
@@ -31,10 +31,10 @@ public slots:
     //
     // These methods raise an exception if they would read
     // beyond the end of the view.
-    qint8 getInt8(quint32 byteOffset);
-    quint8 getUint8(quint32 byteOffset);
-    qint16 getInt16(quint32 byteOffset, bool littleEndian = false);
-    quint16 getUint16(quint32 byteOffset, bool littleEndian = false);
+    qint32 getInt8(quint32 byteOffset);
+    quint32 getUint8(quint32 byteOffset);
+    qint32 getInt16(quint32 byteOffset, bool littleEndian = false);
+    quint32 getUint16(quint32 byteOffset, bool littleEndian = false);
     qint32 getInt32(quint32 byteOffset, bool littleEndian = false);
     quint32 getUint32(quint32 byteOffset, bool littleEndian = false);
     float getFloat32(quint32 byteOffset, bool littleEndian = false);
@@ -51,11 +51,11 @@ public slots:
     //
     // These methods raise an exception if they would write
     // beyond the end of the view.
-    void setInt8(quint32 byteOffset, qint8 value);
-    void setUint8(quint32 byteOffset, quint8 value);
-    void setInt16(quint32 byteOffset, qint16 value, bool littleEndian = false);
-    void setUint16(quint32 byteOffset, quint16 value, bool littleEndian = false);
-    void setInt32(quint32 byteOffset, quint32 value, bool littleEndian = false);
+    void setInt8(quint32 byteOffset, qint32 value);
+    void setUint8(quint32 byteOffset, quint32 value);
+    void setInt16(quint32 byteOffset, qint32 value, bool littleEndian = false);
+    void setUint16(quint32 byteOffset, quint32 value, bool littleEndian = false);
+    void setInt32(quint32 byteOffset, qint32 value, bool littleEndian = false);
     void setUint32(quint32 byteOffset, quint32 value, bool littleEndian = false);
     void setFloat32(quint32 byteOffset, float value, bool littleEndian = false);
     void setFloat64(quint32 byteOffset, double value, bool littleEndian = false);
